remove dependencies to rust feature "naked functions"
- naked functions does currently not work (rust-lang/rust#75922)
- we replaced naked functions by assembly code
- using global_asm to include the assembly code
