+++ src/arch/x86_64/kernel/gdt.rs
@@ -131,6 +131,7 @@ pub fn add_current_core() {
 }
 
 #[inline(never)]
+#[no_mangle]
 pub fn set_current_kernel_stack() {
 	core_scheduler().set_current_kernel_stack();
 }
