+++ src/lib.rs
@@ -28,12 +28,12 @@
 #![feature(allocator_api)]
 #![feature(const_btree_new)]
 #![feature(const_fn)]
+#![feature(global_asm)]
 #![feature(lang_items)]
 #![feature(linkage)]
 #![feature(llvm_asm)]
 #![feature(panic_info_message)]
 #![feature(specialization)]
-#![feature(naked_functions)]
 #![feature(nonnull_slice_from_raw_parts)]
 #![feature(core_intrinsics)]
 #![feature(alloc_error_handler)]
